Trim graph entity shims and split compliance evaluator (#288)
(cherry picked from commit 3104a993f9331c17071084f88701e993ed8e26fb)
