+++ internal/graph/entities/entities_test.go
@@ -36,9 +36,31 @@ func TestDefaultEntityFacetDefinitionsReturnsDefensiveCopy(t *testing.T) {
 	}
 	originalID := defs[0].ID
 	defs[0].ID = "mutated"
+	if len(defs[0].ApplicableKinds) > 0 {
+		defs[0].ApplicableKinds[0] = graph.NodeKindUser
+	}
 
 	again := DefaultEntityFacetDefinitions()
 	if again[0].ID != originalID {
 		t.Fatalf("expected defensive copy, got %q", again[0].ID)
+	}
+	if len(again[0].ApplicableKinds) > 0 && again[0].ApplicableKinds[0] == graph.NodeKindUser {
+		t.Fatal("expected nested facet metadata to be copied defensively")
+	}
+}
+
+func TestEntityFacetAppliesToNodeHonorsApplicableKinds(t *testing.T) {
+	if !EntityFacetAppliesToNode(EntityFacetDefinition{}, graph.NodeKindService) {
+		t.Fatal("expected empty applicable kinds to apply broadly")
+	}
+	def := EntityFacetDefinition{
+		ID:              "bucket_only",
+		ApplicableKinds: []graph.NodeKind{graph.NodeKindBucket},
+	}
+	if !EntityFacetAppliesToNode(def, graph.NodeKindBucket) {
+		t.Fatal("expected facet to apply to listed kind")
+	}
+	if EntityFacetAppliesToNode(def, graph.NodeKindService) {
+		t.Fatal("expected facet not to apply to unlisted kind")
 	}
 }
